requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
ollama>=0.1.0
google-generativeai>=0.3.0
openai>=1.0.0
nltk>=3.8.0
rouge-score>=0.1.2
tqdm>=4.65.0
bert-score>=0.3.13

